worker/requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.34.0
huggingface-hub>=0.27.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
